Absorb snorkel into Fonduer codebase
This removes the snorkel directory and closes #29.
