+++ docs/conf.py
@@ -30,10 +30,8 @@ autodoc_mock_imports = [
     'pandas',
     'pdftotree',
     'psycopyg2',
-    'requests',
     'scipy',
     'spacy',
-    #  'sqlalchemy',
     'tensorflow',
     'treedlib',
     'wand',
