+++ fonduer/__init__.py
@@ -1,22 +1,22 @@
-from fonduer._version import __version__
-
 import logging
 
+from fonduer._version import __version__
 # If we want to directly expose any subclasses at the package level, include
 # them here. For example, this allows the user to write
 # `from fonduer import HTMLPreprocessor` rather than having
 # `from fonduer.parser import HTMLPreprocessor`
-from fonduer.async_annotations import BatchFeatureAnnotator
-from fonduer.async_annotations import BatchLabelAnnotator
-from fonduer.candidates import CandidateExtractor, OmniNgrams, OmniFigures
-from fonduer.matchers import LambdaFunctionFigureMatcher
-from fonduer.models import (Meta, Document, Phrase, Figure, candidate_subclass)
-from fonduer.parser import HTMLPreprocessor
-from fonduer.parser import OmniParser
-from fonduer.snorkel.annotations import load_gold_labels
-from fonduer.snorkel.learning import GenerativeModel, SparseLogisticRegression
-from fonduer.snorkel.matchers import (RegexMatchSpan, DictionaryMatch,
-                                      LambdaFunctionMatcher, Intersect, Union)
+from fonduer.async_annotations import (BatchFeatureAnnotator,
+                                       BatchLabelAnnotator, load_gold_labels)
+from fonduer.candidates import CandidateExtractor, OmniFigures, OmniNgrams
+from fonduer.learning import GenerativeModel, SparseLogisticRegression
+from fonduer.matchers import (DateMatcher, DictionaryMatch, Intersect, Inverse,
+                              LambdaFunctionFigureMatcher,
+                              LambdaFunctionMatcher, LocationMatcher,
+                              MiscMatcher, NumberMatcher, OrganizationMatcher,
+                              PersonMatcher, RegexMatchEach, RegexMatchSpan,
+                              Union)
+from fonduer.models import Document, Figure, Meta, Phrase, candidate_subclass
+from fonduer.parser import HTMLPreprocessor, OmniParser
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
@@ -24,20 +24,28 @@ __all__ = [
     'BatchFeatureAnnotator',
     'BatchLabelAnnotator',
     'CandidateExtractor',
+    'DateMatcher',
     'DictionaryMatch',
     'Document',
     'Figure',
     'GenerativeModel',
     'HTMLPreprocessor',
     'Intersect',
+    'Inverse',
     'LambdaFunctionFigureMatcher',
     'LambdaFunctionMatcher',
+    'LocationMatcher',
     'Meta',
+    'MiscMatcher',
+    'NumberMatcher',
     'OmniFigures',
     'OmniNgrams',
     'OmniParser',
+    'OrganizationMatcher',
     'PDFPreprocessor',
+    'PersonMatcher',
     'Phrase',
+    'RegexMatchEach',
     'RegexMatchSpan',
     'SparseLogisticRegression',
     'Union',
