+++ fonduer/features/table_features.py
@@ -3,7 +3,7 @@ from builtins import range, str
 from fonduer.config import settings
 from fonduer.lf_helpers import (get_cell_ngrams, get_col_ngrams,
                                 get_head_ngrams, get_row_ngrams)
-from fonduer.snorkel.models import TemporarySpan
+from fonduer.models import TemporarySpan
 from fonduer.utils_table import min_col_diff, min_row_diff, num_cols, num_rows
 
 FEAT_PRE = 'TAB_'
