+++ fonduer/learning/disc_models/rnn/re_rnn.py
