+++ fonduer/learning/disc_models/rnn/utils.py
