--- fonduer/snorkel/__init__.py
@@ -1,6 +1,0 @@
-from __future__ import absolute_import
-
-from .models import Meta
-
-__version__ = '0.6.2'
-__all__ = ['Meta']
