--- fonduer/snorkel/learning/__init__.py
@@ -1,25 +1,0 @@
-"""
-Subpackage for Snorkel machine learning modules.
-"""
-from __future__ import absolute_import
-
-from .utils import (reshape_marginals, LabelBalancer, Scorer, MentionScorer,
-                    binary_scores_from_counts, print_scores, GridSearch,
-                    ModelTester, RandomSearch, sparse_abs, candidate_coverage,
-                    LF_coverage, candidate_overlap, LF_overlaps,
-                    candidate_conflict, LF_conflicts, LF_accuracies,
-                    training_set_summary_stats)
-from .disc_models.rnn import reRNN, TextRNN
-from .disc_models.logistic_regression import (LogisticRegression,
-                                              SparseLogisticRegression)
-from .gen_learning import (GenerativeModel, GenerativeModelWeights)
-
-__all__ = [
-    'reshape_marginals', 'LabelBalancer', 'Scorer', 'MentionScorer',
-    'binary_scores_from_counts', 'print_scores', 'GridSearch', 'ModelTester',
-    'RandomSearch', 'sparse_abs', 'candidate_converage', 'LF_coverage',
-    'candidate_overlap', 'LF_overlaps', 'candidate_conflict', 'LF_conflicts',
-    'LF_accuracies', 'training_set_summary_stats', 'reRNN', 'TextRNN',
-    'LogisticRegression', 'SparseLogisticRegression', 'GenerativeModel',
-    'GenerativeModelWeights', 'candidate_coverage'
-]
