--- fonduer/snorkel/parser/__init__.py
@@ -1,18 +1,0 @@
-from __future__ import absolute_import
-
-from .corpus_parser import CorpusParser, CorpusParserUDF
-from .doc_preprocessors import (DocPreprocessor, TSVDocPreprocessor,
-                                TextDocPreprocessor, CSVPathsPreprocessor,
-                                HTMLDocPreprocessor, XMLMultiDocPreprocessor)
-from .parser import Parser, ParserConnection, URLParserConnection
-from .spacy_parser import Spacy
-from .rule_parser import (Tokenizer, RegexTokenizer, SpacyTokenizer,
-                          RuleBasedParser)
-
-__all__ = [
-    'CorpusParser', 'CorpusParserUDF', 'DocPreprocessor', 'TSVDocPreprocessor',
-    'TextDocPreprocessor', 'CSVPathsPreprocessor', 'HTMLDocPreprocessor',
-    'XMLMultiDocPreprocessor', 'Parser', 'ParserConnection',
-    'URLParserConnection', 'Spacy', 'Tokenizer', 'RegexTokenizer',
-    'SpacyTokenizer', 'RuleBasedParser'
-]
