+++ fonduer/utils.py
